+++ db/migrate/20150414161311_add_archived_to_project_favorites.rb
@@ -1,0 +1,5 @@
+class AddArchivedToProjectFavorites < ActiveRecord::Migration
+  def change
+    add_column :project_favorites, :archived, :boolean, null: false, default: false
+  end
+end